Update test routines and documentation
Update the test routines with behaivor for pre-install running of the test
routines.  Check for the existance of the new routines prior to running
the test.

Add a bit to the unit-tests.md file about the known failures on a system
without the patch installed.

Remove comment regarding incorrect documentation about GT.M and this
software.